Remove circular dependancies
